[CodeCompletion] Move CodeCompletionResultSink to its own file
